Include rtype designation to those variables that didn't match BFB with the intel compiler option '-fpconstant'.
As mentioned in the previous commit, there were a handful of locations
where the addition of "_rtype" did not match BFB with the original
code using the '-fpconstant' compiler option.  Those were left out
of the previous commit but are fixed here.

NOTE: one inconsistentcy that we tracked down was that the expressions:
thrd = 1./3.
sxth = 1./6.
remained single precision even with the compiler option.  So it makes
sense that the addition of "_rtype" would change these values and
subsequently all values that use thrd and sxth.  This is likely a bug
or issue with "-fpconstant" and not with p3.