Change optional checkpoint of initialization to mandatory
The addition of a namelist variable and env_run.xml support for enabling
optional checkpointing of performance data during intialization in the
previous commit is backed out here. Instead a new checkpoint timing file is
now output right before the beginning of the run loop, with a name consistent
with the other checkpoint timing files, i.e. including the simulation
timestamp in the suffix. This is not optional.

The internal reworking of the checkpoint write logic and tweaking of timer
names in cime_comp_mod.F90 is preserved, and missing documentation for the
TPROF_TOTAL env_run.xml variable is also added to config_component_e3sm.xml.

[BFB] - Bit-For-Bit