Remove 'xtime' from 'lbc_in' stream
The 'xtime' field was previously being read by the 'lbc_in' stream, but
this is not necessary and in some cases may actually be problematic.
While the 'xtime' field must be present in the LBC files to allow the stream
manager to seek to the correct LBC update file at any given point in
the model simulation, we don't actually need to read the 'xtime' field.
In cases where the time in the LBC file being read does not exactly match
the current simulation time -- because we don't request an exact time match in
the stream manager read calls in mpas_atm_update_bdy_tend -- the xtime field
that was read in would overwrite the model xtime field with an incorrect time.
The 'xtime' field is set to the correct simulation time in the atm_timestep
routine in each timestep, but this happens only after most of the physics
(all non-microphysics schemes, currently) are called.