Delete atm_advance_scalars_work and rename atm_advance_scalars_work_new
In support of limited-area simulations, the atm_advance_scalars_work_new routine
was developed as a replacement for atm_advance_scalars_work. As the latter is
no longer called from anywhere in the dynamics, and as the former works for
global and for limited-area simulations, we can delete the latter and rename
the former to atm_advance_scalars_work.