Set explicit default value of zero for {bdyMask,specZoneMask}x{Cell,Edge,Vertex}
The correctness of the MPAS-Atmosphere dynamics critically depends on several
boundary fields having a default value of zero, and this commit assigns
an explicit default of zero to the following fields in the Registry.xml file:

bdyMask{Cell,Edge,Vertex}
specZoneMask{Cell,Edge,Vertex}

Although the "default default value" for numeric fields is currently zero
in the Registry parser/code generator, this default could in principle change
at some point in the future, so it's safest to assign explicit default values
to fields that we rely on in the solver.

The bdyMask fields are used in the dynamics in code that is always executed
for global and limited-area simulations (e.g., atm_set_smlstep_pert_variables
and atm_recover_large_step_variables), and the mpas_atm_setup_bdy_masks routine
now assumes that the specZoneMask fields are zero by default.