Integrate tridiag solver library, part 2.
ctest unit test for correctness.

Same exe is also a performance driver for comparison with the tridiag nano-app
in scream-docs. Comparison scripts will go in that repo. The objective is to be
sure the performance of the solver when integrated into scream (cmake system,
different header stuff) matches that of the single-file Makefile-driven
nano-app.