Add init_atmosphere option to read vertical coordinate values from a text file
A new namelist option, config_specified_zeta_levels, in the &vertical_grid group
may be set to the name of a text file containing vertical level values, one value
per line. The number of lines in this text file must be one more than the value
of config_nvertlevels, since config_nvertlevels gives the number of vertical
layers, while the zeta values in the text file provide the vertical coordinate
values at layer interfaces.

The config_specified_zeta_levels option is hidden by default, and its default
value is an empty string. If this option is not set to the name of a text file,
the setup of the vertical grid proceeds as in MPAS v6.0, using the value of
config_tc_vertical_grid and config_ztop. If config_specified_zeta_levels is
specified, the vertical levels are taken from the file and config_ztop is
ignored.

The values provided in the vertical level text file must be in increasing
order.