Halt if LBC data cannot be successfully read
The mpas_atm_update_bdy_tend routine, which is responsible for reading in
LBC state and producing LBC tendencies, now returns an error code to indicate
whether data were successfully read from the 'lbc_in' stream. If this routine
does not return a non-zero value, the atm_core_run routine will print an error
message and return with a non-zero status, effectively ending the simulation.

Note that at present, the mpas_stream_mgr_read routine doesn't appear to
return a non-zero error code when we attempt to read from a non-existent
file strictly after the current time. Until this issue is addressed,
the checks introduced in this commit won't correctly stop the model when, e.g.,
the LBC interval is less than the interval of the LBC files provided.