Change in driver's init operations order
Before:
  1. create processes
  2. place fields requests in the repo
  3. create fields and set them in the processes
  4. call 'initialize' on processes, passing the grids manager

After:
  1. create processes
  2. set the grid in the processes, passing the grids manager
  3. place fields requests in the repo
  4. create fields and set them in the processes
  5. call 'initialize' on processes

The reason is that in order to create fields, the processes need to know
the grid's specs. However, the grid was not known until the 'initialize' call.
Hence, the process could not provide valid fields ids until the initialize call.

This was not obvious while working only on the HommeDynamics process,
since that is a special case: Homme is in charge of creating the grids,
so HommeDynamics had the privilege of knowing its grid during construction
time, where Homme is init-ed. To make things clear, I changed the impl
of HommeDynamics too: it now creates fields in the set_grid call, even
if, in theory, it could already to it in the constructor, since it already
has all the required info.

IMPORTANT: the process is *expected* to have valid field ids for
           required/computed fields once it set_grid returns.