Bugfix for variables passed from P3 microphysics to radiation scheme.
This commit fixes an issue where the values passed from the P3-microphysics
scheme to the radiation scheme were inconsistent with P3.  This commit:

1) Removes a number of subroutines taken from MG microphysics that are
no longer needed/used.
2) Uses the effective radius for liquid (rel) and ice (rei) as determined inside P3,
instead of taking values that are calculated based on MG subroutines.
3) Furthermore, the bulk density of ice used to calculate the effective diameter
of ice (dei) is taken from P3 instead of being a fixed constant.