Introduce early check on whether boundary terrain blending will succeed or not
To avoid having the init_atmosphere core stop due to the inability to blend terrain
along the boundary of a limited-area domain only after other time-consuming processing
(e.g., interpolation of static fields) has taken place, a check is introduced at
the beginning of the config_init_case==7 code in init_atm_setup_case. The new check
calls the blend_bdy_terrain routine with a new argument, dryrun, set to .true. to
instruct the routine to simply check whether the intermediate file indicated in
the namelist exists and contains a terrain field. If config_blend_bdy = true in
the namelist and this check fails, we can stop before spending time on any other
processing.