apply the phenology flux limiter
When it is on by setting use_phenolmte to true, this flux limiter
avoids the occurance of negative litter fluxes resulting from
the phenology calculation. This feature is essential when the elm
big leaf model is coupled with plugged-in reactive transport
models such as BeTR. However, since the default elm is resilient to
negative litter fluxes, this feature is by default off.