Avoid calling mpas_atm_update_bdy_tend whenever config_apply_lbcs = false
For global simulations in which config_apply_lbcs is necessarily set to
.false., if a valid input interval is provided for the 'lbc_in' stream,
the input alarm for this stream would ring even though the stream has no
active packages. This caused the mpas_atm_update_bdy_tend routine to be
called, leading to a one of several failures, e.g., a segfault due to
the use of unallocated (due to packages) lbc_* fields in the routine.

In order to avoid this, we now call the mpas_atm_update_bdy_tend routine
only if the 'lbc_in' input alarm is ringing and config_apply_lbcs == true.