Add logic to control activation of glcshelf_c2_ice
Previously there was not an ability to disable the new glcshelf_c2_ice
coupling direction if ice was active.  This commit uses the recently
added ocn_c2_glcshelf flag to control glcshelf_c2_ice (in addition to
the glcshelf_c2_ocn direction it already controls).  The thinking is
that glc-ice coupling is never expected without an active ocean and
glc-ocn coupling.  More refined control could be added later as
situations call for it.

The important result is that glcshelf_c2_ice is default false unless the
ocean model driver explicitly sets the ocn_c2_glcshelf flag to true.

This commit also fixes logic for calling prep_ice_shelf_calc_g2x_ix,
so that it's called based on glcshelf_c2_ice instead of glcshelf_c2_ocn.