Communicate effective radii for cloud, ice, and snow between MPAS and WSM6
The effective radii fields for cloud, ice, and snow (re_cloud, re_ice, and
re_snow) were previously transferred between the rest of MPAS-A and
the Thompson microphysics scheme, but they were not transferred for
the WSM6 microphysics scheme. When config_microp_re = true in the namelist,
the RRTMG radiation schemes would then not have access to the effective radii
from the WSM6 scheme.

This commit moves the communication of these fields to and from microphysics
into code blocks that are executed for both Thompson and WSM6.