Add plant date subroutine for crops
Introduces a new method to calculate crop planting dates. Calculates the
temperature or precipitation seasonality of a grid and determines planting
date from temperature or precipitation thresholds. Therefore, the growing
season is either the warm or wet season. This allows crops to grow globally,
rather than in just the temperate latitudes.

A new module, CropMod is added to support some of the calculations and
will serve as a home of future crop model developments that are specific
to crops. The parameter file is not changed as there are no new parameters.

[Non-BFB] - Non Bit-For-Bit (only for the crop model)

See confluence for a more detailed description about these tags.