Removed initialization of dp3d
Instead of making sure dp3d was initialized, it was pointed out during the PR review that dp3d cannot be trusted and may potentially be removed from the element structure in the future. So as a safer alternative the dp variable needed for mapping to the FV physics grid was simply calculated in fv_physics_coupling_mod from the surface pressure on the dynamics grid.

	modified:   components/cam/src/dynamics/se/dyn_comp.F90
	modified:   components/cam/src/dynamics/se/fv_physics_coupling_mod.F90