Adding Land Use Translator code
The Land Use Translator (LUT) converts either historical or future land-use data (in Land-Use Harmonization 1 format) into land cover changes in terms of 16 different plant functional types. The output of the LUT is used by mksurfdata for generating surface files for E3SM. This version of the LUT does not directly interact with the E3SM code - it is just a tool that generates inputs for E3SM. This version of the LUT is modified from the version used for iESM and from the original version created by Peter Lawrence, although it still retains a significant amount of code from those versions.

 On branch lchini/clm/landusetranslator
 Changes to be committed:
	new file:   full_updateannuallanduse_louise_cleanedup.c
[BFB]
EB-36, EB-38