Reset pointer to mpas_log_info at the start of each call to mpas_{run,finalize}
Because the mpas_log module maintains a pointer to the currently active logging
info as a module variable, this pointer must be set to the logging info for
the currently active domain at the beginning of each call to mpas_run and
mpas_finalize.

There is no need to perform similar logic in mpas_init, since the call to
the core % setup_log() function call invokes mpas_log_init(), which internally
sets the logging info pointer.