Put back in a limiter on qidep and qisub that was removed.
There was a limiter on qidep and qisub that was removed because we
thought it was related to the super-saturation adjustment that was
not being removed from P3.  This allows qidep to become too large
causing an instability to form.

This commit reinserts the original limiter on qidep and qisub to
remedy this issue.