Adding functionalities to AtmosphereProcessGroup, plus minor changes
The APG is now in charge of creating AbstractRemapper object to remap
each AP inputs/outputs from/to the reference grid to/from the AP's grid.

These remaps are done automatically before/after the call to the AP's run
method, so no additional code is required by the AP's developers.

Also, added capability to store two copies of the field repository in the
APG. These field repos are used to make sure that the stored AP's are not
actually changing fields that they did not declare as computed. We do this
by having two hard copies of the field repo, and comparing all the views
that are not declared as computed in the two repos. They should be identical.
This check is only performed if SCREAM_DEBUG is defined, and if the two
field repos have been set in the class.