Modify gather_scatter routines to use smaller MPI message tags
The routines in ice_gather_scatter.F90 use message tags that
are offsets of the global block id. For very high resolution runs
using SE meshes, e.g. ne1024, the number of blocks can be very
large, exceeding the maximum MPI message tag on some of our
targer systems, e.g Cori-KNL. The use of global block ids is
not necessary, and here the algorithms are restructured to
eliminate this dependency, depending instead on the number
of blocks on each process after the decomposition. Even this
may not be sufficient unless a large number of processes
is used, but it does provide a mechanism to decrease the
message tag size, and also save some memory.

BFB