Allow for running ctests in parallel with 'ctest -jN'
Our test-generation marcro already knew about threads and mpi-ranks,
so only a simple multiplication was needed to get the core count needed
for a test. Using this value to set the PROCESSORS test property is all
that was needed to make good use of Ctest's -jN flag, which gives the user
a "pool" of N cores with which to run tests.

On melvin, a large desktop, using ctest -j48 reduced the time needed to run
all screams test from about 90 seconds to about 10 seconds.

Other changes:
1) Make every test, include p3_regression (now called p3_run_and_cmp) go through
   our CreateUnitTest macro.
2) Reduce the cost of workspace testing. Soak tests are off by default.
3) Reduce the cost of lin-interp testing. Tests the use thread-sweeping should be
   very fast to run, otherwise it bogs down testing.
4) Replace the NAMELIST_FILE arg to CreateUnitTest with a more-generate EXE_ARGS.
5) Allow travis to do testing with -j2
6) Use the CTEST_PARALLEL_LEVEL environment variable as the -j for building for
   our core ctest testing script.
7) Add DEP argument to CreateUnitTest to allow for test dependencies
8) Express that all p3 tests dep on the first p3 test (in order to get .dat8 files)