Bug fix in the AD
Always request that the grids manager build the reference grid. Usually,
the ref grid is one of the grids used by one of the atm processes. However,
if by chance (perhaps during testing) the user asks for a ref grid that is
not used by any of the atm processes, the gm should still build it, to avoid
crahses at run time.