Adds AODALL to compute total aerosol (incl. nighttime) optical depth
Aerosol optical depth (AOD) and column burden are currently calculated
in the model during daytime only, which makes the burden output
little useful. This module is modified to calculate burden and total
 optical depth during nighttime as well. The burden diagnostic
variables already exist in the model but a new variable AODALL is
added in order to keep the original one (AODVIS). The 2-D burden
variables for each size mode (BURDEN1, BURDEN2, …), four of them for
the default configuration, are moved out of the "history_verbose" if
loop, so that they can be written out when history_aero_optics is
set to "true".

[BFB] - Bit-For-Bit