Homme: Changed random initialization of ElementsXYZ/Tracers structures
The structures need to be inited first, and only in a second stage they can
be randomized. This allows to call the 'randomize' method multiple times,
without re-initing the views (which would change the pointers stored, which
could be a problem in cxx-vs-f90 testing).

Also, changed the HVF constructor, to take ElementsGeometry, ElementsState,
and ElementsDerivedState separately, rather than the big Elements structure.