A small upgrade to the catch-based testing system
It is now possible to use the `--scream-test-params <key=var,...>`
syntax to pass one or more options to the particular test. It is up to
the individual test whether that option is valid and/or used. A test
is free to ignore the options altogether. All the pairs key-value are
stored as string, so if the test needs, say, an int, it must convert
the string to an int.

Example usage: my_test --scream-test-params ne=2,qsize=4,moist=false

Fixes #140