Revert to old oRRS15to5 domain
Revert to older oRRS15to5 ocean domain file. I had changed this to be
consistent with the newly generated domain files (that used new maps),
but since this file actually does not depend on the maps (ocean
coordinates copied straight through) this change was not needed. So,
reverting to minimize changes here.