Cross-language/architecture BFB diagonally dominant tridiag solvers.
Add scream::tridiag::bfb and Fortran-callable equivalents. These will be used
when doing BFB conversion of SHOC and of HOMME theta-l NH. Obviously they are
not performant.

These are tested for correctness in the usual property test. Then a new BFB test
verifies F90 and C++ on device agree BFB.