HOMME: Remove scans from Newton loop.
In the performance-portable code, we want to avoid scans, as these are
costly on GPU. The residual function F is fundamentally a function of
phi, not dphi, so we can remove all scans from the routine. Instead,
we update phi_np1 as the Newton iteration proceeds and recompute dphi
(which requires just a simple parallel_for instead of parallel_scan)
from phi_np1.