Upgrade to test-all-scream
In case of multiple builds to test (dbg, fpe, sp), the script can
now launch the builds in parallel. This can help if the compilation
or testing phase of a build reaches a point where 1-2 files or tests
are taking much longer than the others, leaving most of the machine idle.

Most importantly, the different builds are now performed in separate
folders, whose names incorporate the build name. This allows to launch
multiple instances of test-all-scream in the same folder, with different
builds, which can then run in parallel.