Move heatDissipation and basalFrictionFlux calculations
These are moved from the thermal driver to the velocity driver.  Then
these fields are made restart variables, and not calculated on the
initial timestep if the run is a restart.

This is required to have those fields correct on a restart, because many
of the fields that are input to their calculation are not restart
variables.

With this change the two new thermal restart tests pass.