Added generic-desktop-serial machine file
On some machine (e.g., blake), I was getting an error
looking for OpenMP when using gcc. After googling a bit,
this error happened to others (in Trilinos too), which
ended up using hacks on CMAKE_XYZ_FLAGS in order to
bypass it. I simply decided to have a serial machine file.