Fix erroneous memset calls
Two calls to memset in gptl.c in the timing library have their
arguments in the wrong order. The other calls to memset in gptl.c are
correct. In practice this has not caused any problems (due to the
values being swapped), and the result is BFB. To avoid future issues,
it is still important to fix this.

[BFB]