Homme: minor improvement to genRandArray
When a constraint is passed, try only a limited number of times.
If the constraint is not met in those attempts, it may be a sign
that the constraint and the pdf are not compatible.

E.g., pdf is U(-1.0,0.0), but the constraint checks that the
generated values are positive.