Homme: bug fixes related to unit testing of RemapFunctor
It all sparked from the check that m_pio and m_pin in PpmRemap should be
"close" at the surface (in fact, they should match, but the check only
made sure they were less than 1.0 apart). These pressures are computed
as integral of thickness (old and new, aka source and target).

In the case rsplit=0, the src thickness is a perturbation of the tgt one,
with perturbation equal to dt*(eta_dot_dpdn[k+1]-eta_dot_dpdn[k]).
This imposes a constraint to eta_dot_dpdn too. In particular, all is good
if the sum of delta(eta_dot_dpdn)=0 over a column. To do this, we simply
subtract sum(delta(eta_dot_dpdn)) from its first entry after we randomly
generate it. This is not 'exact', but is ok up to fp roundoffs.

Meanwhile, I also replaced some kernels in RemapFunctor with existing
kernels from HybridVCoord or ColumnOps, to reduce code duplication.