cube_to_target: Prep this utility for physgrid work.
In the new physgrid topography toolchain, we need to run this utility twice. On
ne30pg2 (for example), it would previously issue a message and abort. However,
with the stop statement removed, the resulting topo dataset appears fine. So
comment it out.

Abort immediately if --smooth-topography is an invalid file.

Reduce the number of progress messages.