Add irtc_rate variable to list of private variables in threaded region
The variable irtc_rate was mistakenly left out of the private variable
declarations in the OMP directives implementing threading over chunks
in three locations in physpkg.F90. This probably contributes to the
error observed when calculating chunk costs with OpenMP threading
(but fixing this is not sufficient in itself - adding thread safety to
the calls to shr_sys_irtc is also required). Here the irtc_rate is
added to the necessary private variable declarations.

[BFB]