Remove unnecessary inclusion of optional shr_sys_irtc parameter
The utility function shr_sys_irtc has an optional parameter called
'rate' in which the system clock resolution is returned. This
parameter is included in all calls to shr_sys_irtc in physpkg.F90, but
is only needed in half of these calls. Here this is remedied, and the
location of the associated local variable, irtc_rate, is moved in the
OpenMP PRIVATE clause to better reflect its location in the subsequent
code block.

[BFB]