Document when shr_sys_irtc can be used in multi-threaded region
The addition of the threadprivate declaration to shr_sys_irtc is
sufficient to ensure thread safety as shr_sys_irtc is used within
E3SM currently. It is not sufficient for all possible usage scenarios,
e.g., between multi-threaded regions when the two regions have
different numbers of active threads. Comments are added to
shr_sys_irtc to document this.

[BFB]