Disable call to shr_taskmap_write in OCN when INFO_TASKMAP_COMP is zero
If INFO_TASKMAP_COMP in env_run.xml is zero (the default), then the mapping
of MPI tasks to cores is not output, but it is still calculated. For
large task counts, this can be expensive, and it is not necessary for
MPAS-OCEAN. (Note that the calculation is still necessary for CAM,
even when the mapping is not output.) Here this is addressed, and
shr_taskmap_write is no longer called for MPAS-OCEAN when
INFO_TASKMAP_COMP is zero.

[BFB]