Add variables to crmout test
This adds another cam history file to test a range of output variables
on the global grid. If this is compared with an old crmout test there
will be a failure because of the missing file, but shouldnt really
affect the test otherwise since the CRM grid output will still be
compared.

	modified:   components/cam/cime_config/testdefs/testmods_dirs/cam/crmout/user_nl_cam