Implementation to use cosp2 as submodule
COSP2 is introduced as a replacement for the existing COSP.
The core cosp2 simulator codes are imported as a submodule from
git@github.com:CFMIP/COSPv2.0.git. Its branch of CESM_v2.1.4 is
currently used. The interface module for cosp as well as the supporting
cosp modules that depend on a model's physical parameterization (e.g.,
species considered in cloud microphysical parameterization) are
implemented in E3SM repository.

MODIS simulator is currently turned off because of a known issue yet to
be resolved.

[BFB] except for COSP diagnostics