Hommexx: upgrade to gradient_sphere
As it was done for divergence_sphere, we now allow
the input field to be provided even in functional form
(rather than by a view). The input field provider must
be implement

  Scalar operator() (int, int, int)

which returns the value of the field at given GP pair,
and given level.

This allows the user to pass a lambda to compute the
input field on the fly, rather than pre-compute it and
store it.