Set spectral bounds based on RRTMGP inputdata
Set spectral band limits based on values from RRTMGP inputdata files,
rather than hard-coding these values. This makes the interface more
flexible and helps avoid confusion as to what bands are in use. Also add
a routine to check that these band limits are set correctly. This
routine is to be used for testing purposes, and is not called during
runtime by default, but a call can be inserted after setting the band
limits to effectively unit test to confirm that bands are set consistent
with what we expect.