Changed the use of sd_cs in call to subsample_and_optics to fix runtime failure in debug mode.
The runtime failure in debug mode is described in issue #3482.

Allocatable sd_cs(lchnk) is used as real argument in call to subsample_and_optics.
But sd_cs is only allocated when lradar_sim is true. When lradar_sim is
false as in cosplite test, it is unallocated, while the dummy argument
asks for one that is either allocated or not an allocatable variable.

It is change to use an intermediate variable sd_wk, which takes value
from sd_cs(chnk) when lradar_sim is true before calling
subsample_and_optics and assign back after the call.

[BFB] for this change alone, but fix the cam-costlite_nhtfrq5 test.