Detect MPI from a conda environment setup_testcase.py
Add support for a "conda_mpi" attribute to "step" tags. If this
attribute is set to "true" and MPI is present in the conda
environment, that command will be called with `mpirun` from
the conda envrionment.

This is needed to support compass conda enviornments with mpich.
Python scripts and modules that use the netcdf4 package with
mpich support don't work properly on many compute nodes (e.g.
Grizzly at LANL and Anvil at ANL) unless they are prefixed with
`mpirun -np 1`