Add interface extension map for marine grounded margine
This map is needed to extend observational fields at grounded marine margins.
A second boundary extension map is created for these boundaries.  It is then
used in the function that imports the observational fields to copy the obs values
at the last location with obs into these extensions.

Also, a minor bug fix is included that ensures that at the 2m ice shelf appended
at marine grounded margins the beta value is set to 0 under the floating ice.