Switch algorithm for finding layer thicknesses
This is a more straight-forward root finding approach using
scipy.  It finds a characteristic lengh scale `delta` over which
layer thickenesses vary from `dz1` a the surface to `dz2` at
`z --> -infinity` such that the sum of the layer thicknesses
hits the target bottom depth in the given number of layers.

Docstrings and formatting have been improved.