CIME: the buildlib method in buildlib_cmake files can now return a string
If buildlib returns a string, it will contain cmake flags to be used
for building the current lib or component.
In cime's build.py we then append those flags to the current cmake
command line (if building with cmake).

This allows each component to do all its specific cmake stuff
in their buildlib_cmake file, so that build.py can stay
component-agnostic.