Physgrid: Clean up some geometry calculations.
* Make area calculation more accurate:

It was reproted that the sphere area on the physics grid was inaccurate and
triggering a check in CAM. This is benign because the area used for mass
conservation agrees at the element level with the dynamics grid to machine
precision. But it still needs to be fixed. I saw this:

ne1024
  0: gfr> init nphys  2 check T ftype 0 boost_pg1 F
  0:  gfr> area fv raw   12.5663706169810       2.086357642668703E-010 <- bad
  0:  gfr> area gll      12.5663706143622       2.431357356496555E-013 <- ok

With this fix, I now get:

ne512
   0: gfr> init nphys  2 check T ftype 0 boost_pg1 F
   0:  gfr> area fv raw   12.5663706143631       3.088671990665682E-013
   0:  gfr> area gll      12.5663706143605       1.030499716794179E-013
ne1024
   0: gfr> init nphys  2 check T ftype 0 boost_pg1 F
   0:  gfr> area fv raw   12.5663706143498       7.458047333067340E-013
   0:  gfr> area gll      12.5663706143622       2.431357356496555E-013
ne2048
   0: gfr> init nphys  2 check T ftype 0 boost_pg1 F
   0:  gfr> area fv raw   12.5663706143177       3.299012673599803E-012
   0:  gfr> area gll      12.5663706143428       1.299221247881386E-012
ne3072
   0: gfr> init nphys  2 check T ftype 0 boost_pg1 F
   0:  gfr> area fv raw   12.5663706135615       6.347736897466303E-011
   0:  gfr> area gll      12.5663706144224       5.028951704344270E-012

* Move metric tensor to cubed_sphere_map=2 center: Previously, the metric tensor
used in remapping vector quantities between grids was computed at the reference
element center. But cubed_sphere_map=2 defines the center as the average of the
four element corner points projected to the sphere. Fix this.