SCREAM: created AtmProcDAG class, and added a test.
The class can create a dag output file, to be fed to graphviz. The current
test does nothing other than make sure it runs.

We also added a dummy remapper for SE grids. It does no actual remapping,
but it's useful for setup operations that require a valid remap to be set.
See doc in share/tests/dummy_se_grid_remapper.hpp for more info.