SCREAM: started to work on Field initialization framework
The FieldInitializer will be a structure that the AD uses to init those
fields that are 'input' to the whole atmosphere, during initialization.

The AD will also check that for each field that is an atm input there is
an init policy set, and, if the policy is to use a FieldInitializer,
that a valid FieldInitializer has been registered for the field.