Allow user to specify an extra machines directory in create_newcase
This is to support making an on-the-fly cime port (see ESMCI/cime#3493
for more rationale).

Note that this extra machines directory is used in addition to the
~/.cime directory.

I tested this on my mac (which uses porting information in ~/.cime), by
first creating a case and running case.setup as normal from cime master;
then moving ~/.cime to a temporary location, doing the same
create_newcase from this branch but with --extra-machines-dir pointing
to the temporary location of my .cime directory, running .case.setup in
that new case, and comparing the resulting Macros.make and Macros.cmake
with the baseline. They were identical, as hoped for.

Resolves ESMCI/cime#3493