Homme: Rewrite VR's linextrap for BFB with -qopenmp.
I noticed a particular non-BFB situation: 1 thread with and without
-qopenmp. With -qopenmp, everything is BFB regardless of number of
threads. So -fp-model fast -O3 -qopenmp was producing different code
than -fp-model fast -O3. I believe there is no E3SM test that detects
this kind of non-BFBness.

I traced at least once instance of this non-BFBness to linextrap in
vertremap_base.F90. This commit rewrites linextrap to remove this
non-BFBness when tested with Intel 18.0.1. The new version is the same
as the old in exact arithmetic but not in FP.