Change F-Compsets for super-fast chemistry so that ieflx_opt = 0
The compsets in the previous commit for the superfast chemistry used
the same use-case files for both the B and F compsets. However, the
energy fixer should be set differently for B and F compsets.  This
commit fixes this by creating separate use-case files for the
F-compset, for which the only change is that

   ieflx_opt = 0

In making this change, the compsets

   F1850-AR5-superfast
   F20TR-AR5-superfast

have been replaced by

   F1850S-AR5-superfast
   F20TRS-AR5-superfast

[BFB, except for the superfast compsets]