Cleanup of p3_interface.F90
Phase 1: Remove all unused variables and code.

This commit removes commented out lines that were left in just in
case we decided we needed them.
It also removes a series of variables that are declared but not used,
and in a few cases are treated as output from p3_main but not used.

Finally, a minor bug where numliq is the appropriate variable to use
instead of nc for a diagnostic field.

B4B