Continued cleanup of micro_p3_interface.F90
Phase 2: Remove all unused PBUF variables and remove any reference
to sub-columns (which will not be used in SCREAM-P3).

This commit audited the set of PBUF variables accessed in the P3
interface and removed those that are unused.

This commit also removed all reference to sub-columns, which included
replacing the variable ngrdcol with ncol, since ngrdcol=ncol unless
sub-columns are in use.

BFB