Fix zero-size sw_band_midpoints and lw_band_midpoints output
Fix bug that was causing us to try to write sw_band_midpoints and
lw_band_midpoints using zero-size arrays in the MMF version of the
radiation_init routine. This was caused by these arrays being automatic
arrays that relied on the nswbands and nlwbands dimension sizes having
not been defined yet. Oddly enough, the switch to PIO2 was what caught
this bug, and threw an error at runtime when trying to output these
arrays. Making these arrays allocatable and allocating them with sizes
that are defined earlier in the routine fixes the runtime fails when
using PIO2.