Homme: Fast F90 DIRK solver.
Bacpkort the fast DIRK solver from Hommexx (C++) to Homme (F90). First, since
the Jacobian is strictly diagonally dominant, pivoting is not needed for
stability. Second, since pivoting isn't needed, it's natural to batch the solves
in the GLL direction. This leads to perfect vectorization of the tridiagonal
system solve. Third, since the solves are batched along the GLL direction,
everything in the driver loop can be done in Homme's usual fast layout.

At this commit, with MKLSOLVE defined, this branch is BFB with master. With
MKLSOLVE undefined and so the new solver used, we don't expect BFBness.