Setting the kind for reals in lin interpolation
The kind is set depending on EKAT precision.

Note: setting the flag for default real and default double would also
fix this, but we should not rely on that flag. Ideally, EKAT should not
need that flag, and fully rely on EKAT_DOUBLE_PRECISION to establish
the kind of every real.