Eliminate compile-time dependence on pcols in unicon_cam
The data type unicon_out_t in the unicon_cam module, in
physics/cam/unicon_cam.F90, refers to pcols in dimensioning some
of the fields in the type. These fields are changed to allocatable,
and these fields for variables of type unicon_out_t are then
allocated in unicon_cam_tend, and deallocated in a new subroutine
unicon_cam_tend_free. A call to unicon_cam_tend_free is added to
subroutine convect_shallow_tend in convect_shallow.F90.

This removes one dependency on the compile-time specification
of pcols.

[BFB]