Eliminate compile-time dependence on pcols in phys_grid.F90
The data types chunk and lchunk in the phys_grid module,
in physics/cam/phys_grid.F90, refer to pcols in dimensioning some of
the fields in these types. These fields are changed to allocatable.
Then these fields for variables of type chunk are allocated
in the subroutine phys_grid_init and in the subroutine create_chunks,
and these fields for variables of type lchunk are allocated in the
subroutine phys_grid_init. They are never deallocated as the associated
type chunk and lchunk variable are module variables persistent
throughout the execution of EAM.

Also, the local variable gcols in phys_grid_init is dimensioned as
size pcols. As phys_grid_init is a potential location for optimizing
the value of pcols, this is changed to be allocatable, and then
allocated and deallocated in phys_grid_init.

This removes dependencies on the compile-time specification
of pcols, and also introduces the flexibility of computing an
optimal choice of pcols within physi_grid_init.

[BFB]