Enable compile-time setting of pcols via CAM_CONFIG_OPTS
Currently, adding the -pcols option to CAM_CONFIG_OPTS in
env_build.xml sets PCOLS, which is then used as the compile-time
parameter if -DPPCOLS is also added to Macros.cmake and as the
default value of phys_chnk_fdim otherwise. Here we change the
behavior to force the addition of -DPPCOLS whenever -pcols
is included in CAM_CONFIG_OPTS. This is a more convenient
mechanism for restoring the compile-time setting of PCOLS.
This convenience comes at the cost of removing the ability to
change the default value for phys_chnk_fdim. This capability
is unlikely to be missed, but we can revisit the decision in
the future if it ends up that there is a demand for this
capability.

[BFB]