Add limiters for RRTMGP optics
Add code to support turning on limiters on RRTMGP optical properties.
These are enabled by setting `rrtmgp_clip_optics = .true.` in the
atmosphere namelist (i.e., `user_nl_cam` in a case directory). When
enabled, the `clip_values` function is called to make sure optics going
into RRTMGP are within the range expected by the RRTMGP codes (`0 <= tau
< huge(tau)`, `0 <= ssa <= 1`, `-1 <= asm <= 1`). Note that turning this
flag on clips values and issues a warning, but allows the simulation to
continue. This is to work around the case where values are *slightly*
outside the expected range (due to, e.g., floating point precision
issues), but not that turning this feature on will prevent the radiation
code from catching the much larger errors. Since a warning is spit out
to the log files in this case though, once can simply check the log to
see if optical properties fell well outside of the expected range.