Reorder init. logic for EUL, FV, and SLD dycores
The reordering of the initialization also requires changes
to dyn_comp.F90, inital.F90, and restart_dynamics.F90 for dycores
EUL and FV, and to inital.F90, and restart_dynamics.F90 for dycore SLD.
These are implemented here, but are not tested (as they are not supported
by E3SM). While not testable, the changes are made to avoid future
confusions.

[BFB]