Fix setting of swband/lwband for history files
Fix setting of swband and lwband coordinate variables for history files.
Since add_hist_coord simply sets up pointers to the arrays that are
passed in, the data needs to persist. Thus, we make these arrays
module-scope variables in radiation.F90. Also a little bit of clean-up
in radiation.F90 and radconstants.F90. In particular, we now use
nswbands and nlwbands from radconstants.F90, rather than defining new
versions of these in radiation.F90. Since these values are set at
compile time and should never change, we should be using these in
radiation.F90 anyways, and it does not make sense to make new versions
of these that could inadvertently get changed from what we expect in
radconstants.F90. So instead of setting new values for these based on
the absorption coefficient data, we simply verify that the number of
bands in the input files matches what we expect in the code (14 for,
shortwave, 16 for longwave). These fixes are BFB in regards to the
simulation, but EAM history files will be non-BFB due to the fix in the
swband and lwband coordinate variables. This will only affect
simulations that are long enough to generate history files, i.e., either
long (month-plus) runs, or high temporal output frequency runs (like the
crmout test).