E3SM: make sure CMAKE_BUILD_TYPE is set at the toplevel CMakeLists.txt
As explained in detail in components/CMakeLists.txt, it is not good to
locally change the value of CMAKE_BUILD_TYPE (like scream used to to),
since it may cause very weird config/build errors, VERY hard to track.
Mostly, it boils down to different parts of the project looking for
a different target/variable than the one you would expect.
So, we set it at the beginning of the project (if not already set).

Since CIME only has a bool var (DEBUG), we support only "DEBUG"
and "RELEASE" as possible build types.

However, doing this we allow CMake to "append" flags at the end of
the list of flags for each target. For "RELEASE", this would mean
adding "-O3", while for "DEBUG" it'd mean adding "-g".
While these *may* be reasonable, we do not want to trump flags
that specific components might set, so we also set the variables
CMAKE_<LANG>_FLAGS_RELEASE and CMAKE_<LANG>_FLAGS_DEBUG to empty.

As a side consequence: there no longer is any need to include Macros.cmake
in build_scream.cmake.