Clean up mpas_pool_destroy_pool
This commit first removes the source of a potential deallocation error:

   "mpas_pool_routines.F", line 358: 1525-109 Error encountered while attempting to deallocate a data object.  The program will stop.

when using the XL compilers; this error is a result of the fact that
the mpas_pool_add_subpool routine declares the subpool dummy argument with
the target attribute, and deallocation can only happen for variables that
are allocatable or have the pointer attribute.

With the change to the mpas_pool_add_subpool routine, all "stat" optional
arguments can be removed from calls to deallocate in the mpas_pool_destroy_pool
routine.