Add variables to control new runtime default pcols calculation
The new default runtime pcols that takes into account details of a
particular case, which will be specified in a subsequent commit,
takes into account some new namelist variables:

phys_chnk_fdim_max:  an upper bound on the declared first dimension
                     for physics variables (chunks), for when
                     phys_chnk_fdim < 1.
phys_chnk_fdim_mult: a restriction on the declared first dimension for
                     physics variables (chunks) to be a multiple of this
                     value, for when phys_chnk_fdim < 1.

(Here, when phys_chnk_fdim < 1, the default pcols value for each virtual
SMP is calculated based on the case particulars.)

These can be set via user_nl_cam, but empirically determined default
choices can also be specified in namelist_defaults_cam.xml. An example of
this for phys_chnk_fdim_max for dycore 'se' and phys_grid option 2 is
provided in this commit. Another example, but for a specific system (compy)
as well as for the dycore and phys_grid option, is also provided. Finally,
an example for phys_chnk_fdim_mul for system 'cori-knl' and dycore 'se'
is provided. To support system-specific defaults, a new parameter
'mach' is added to cam/bld/configure .

[BFB]