SCREAM: storing names of generated baselines in baseline dir
This is helpful in a scenario like this:

1) We generate all baselines.
2) Times passes, and we regenerate baselines, but NOT for all of them.
   This can happen if we run test-all-scream with -t flag, and specify
   only some of the possible builds.
   Notice that this does not erase baselines for other builds.
3) We later run test-all using those pre-existing baslines.

In 3), test-all would think all baselines are good, since a) the sha
is the correct one (assume nothing happened in origin/master between
2 and 3), and it can find a baseline data file for all builds.
HOWEVER, the baseline files for builds not enabled in step 2) would
actually correspond to an OLD version of master.

Having a file that stores the name of the baselines that were last
generated is helpful, so we can see if a baseline was not re-generated
last time we generated baselines.