Better handling of Kokkos_ENABLE_Debug
The variable was not being referred to by the correct name in Kokkos.cmake.
Also, this setting did not belong in the machine files since both release and
debug builds are valid for any machine. Instead, we allow the user to select
Kokkos_ENABLE_Debug explicitly; if they do not, then we select a default based
on the build type.