SCREAM: prefer passing C-style strings to Fortran than C++ strings
Even though C++ strings are likely always going to store the char* first
internally (hence, when calling c_f_pointer in F90 we would indeed get
a pointer to the beginning of the string), it is a subtle assumption.
Since what we really want to pass to Fortran is the char* stored
in the string, let's be explicit, and simply pass that.