Improve load balancing when processes have different numbers of threads
In the case when processes do not all have the same number of threads,
the logic in assign_threads may not calculate the number of chunks to
be assigned to each process in a load-balanced way when the total
number of chunks is not evenly divisible by the total number of
threads.

Here, the logic is modified to assign "extra" chunks:

  mod(total number of chunks,total number of threads)

in a round-robin fashion, first to thread 0 in each process, then to
thread 1, etc. If a process does not have a particular thread id, then
it is skipped in this assignment. This approximately load balances
across the threads, which is a more accurate metric when load
balancing performance.

[BFB]