Extension of the scream-scorpio interface to include input.
This commit extends the framework already in place for accessing
scorpio output routines to now include access to scorpio input routines.

The set of unit tests is increased by one, thus a file is created and
has output written to it in test #1.

Test #2 opens that file and reads all fields in as input, testing against
the expected field values.

The testing has also been extended to include up to 4 threads.

[BFB]