Eliminate unuseful warning when specifying pcols at compile-time
There is currently a check whether a runtime pcols specification
(phys_chnk_fdim) disagrees with a compile-time specification. If so, a
warning message is output and the model continues running using the
compile-time specification. This is even triggered by the default
setting of phys_chnk_fdim, which means that the warning message is
always output when pcols is specified at compile-time.

The default (phys_chnk_fdim == 0) implies that the model should
determine what pcols should be set to. When defining pcols at
compile-time, that changes what the model will determine pcols to be
(at runtime), so phys_chnk_fdim == 0 does not conflict with a
compile-time setting of pcols (with this interpretation). Here the
logic is changed to remove the warning when phys_chnk_fdim <= 0 and
pcols is also specified at compile-time.

[BFB]