Update mpas_kd_tree.F to break ties on equidistant queries
Previous to this commit if a query is equidistant from two points of the
KD-Tree, the resulting 'closest' point (from mpas_kd_search) would be the first
point that the KD-Tree visited in its search. This produces different results
across different trees that contain the same ties, but have a
different tree structure.

This commit adds a function to break these ties, so that different trees will
predictively select the same point on equidistant queries.